+++ jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  transform: {
-    "\\.m?jsx?$": "babel-jest"
-  },
-  transformIgnorePatterns: [`/node_modules/`, `/_scratch/`]
-};